Create Extra Q
Get the three angles of a triangle as input.

find the count of the type of the triangle.

Continue the process for 5 times.

If the sum of the three angles is greater than 180 then prompt for correct values. (the sum of all internal angles of a triangle is always equal to 180°). Keep the count of the wrong entries also.

Acute Angled Triangle (all three angles less than 90°)

Right-Angled Triangle (one angle that measures exactly 90°)

Obtuse Angled Triangle (one angle that measures more than 90°)